Replaces PremadeEvent with BentoBoxEvent to make it look nicer.
